Start file split
