+++ src/DCCEXProtocol.cpp
@@ -1795,245 +1795,3 @@ bool DCCEXProtocol::splitFunctions(char *cmd) {
     char* Consist::getConsistName() {
         return consistName;
     }
-
-// class Turnout
-
-    // Turnout::Turnout(int id, char* name, TurnoutStates state) {
-    Turnout::Turnout(int id, TurnoutStates state) {
-        turnoutId = id;
-        turnoutState = state;
-        hasReceivedDetail = false;
-        turnoutName = nullptr;
-        // char *dynName;
-        // dynName = (char *) malloc(strlen(name)+1);
-        // // strcpy(dynName, name);
-        // sprintf(dynName,"%s",name);
-        // turnoutName = dynName;
-    }
-    bool Turnout::throwTurnout() {
-        setTurnoutState(TurnoutThrow);
-        return true;
-    }
-    bool Turnout::closeTurnout() {
-        setTurnoutState(TurnoutClose);
-        return true;
-    }
-    bool Turnout::toggleTurnout() {
-        setTurnoutState(TurnoutToggle);
-        return true;
-    }
-    bool Turnout::setTurnoutState(TurnoutStates action) {
-        TurnoutStates newState = action;
-        if (action == TurnoutToggle) {
-            if (turnoutState == TurnoutClosed ) {
-                newState = TurnoutThrown;
-            } else { // Thrown or Inconsistant
-                newState = TurnoutClosed;
-            }
-        }
-        if (newState<=TurnoutThrow) { // ignore TurnoutExamine
-            turnoutState = newState;
-            // sendTurnoutAction(turnoutId, newState)
-            return true;
-        }
-        return false;
-    }
-    bool Turnout::setTurnoutId(int id) {
-        turnoutId = id;
-        return true;
-    }
-    int Turnout::getTurnoutId() {
-        return turnoutId;
-    }
-    bool Turnout::setTurnoutName(char* name) {
-        // if (turnoutName != nullptr) {
-        //     free(turnoutName);
-        //     turnoutName=nullptr; 
-        // }
-        // char *dynName;
-        // dynName = (char *) malloc(strlen(name)+1);
-        // // strcpy(dynName, name);
-        // sprintf(dynName,"%s",name);
-
-        // turnoutName = dynName;
-        turnoutName = name;
-        return true;
-    }
-    char* Turnout::getTurnoutName() {
-        return turnoutName;
-    }
-    TurnoutStates Turnout::getTurnoutState() {
-        return turnoutState;
-    }
-    void Turnout::setHasReceivedDetails() {
-        hasReceivedDetail = true;
-    }
-    bool Turnout::getHasReceivedDetails() {
-        return hasReceivedDetail;
-    }
-
-// class Route
-
-    // Route::Route(int id, char* name) {
-    Route::Route(int id) {
-        routeId = id;
-        hasReceivedDetail = false;
-
-        // char *dynName;
-        // dynName = (char *) malloc(strlen(name)+1);
-        // // strcpy(dynName, name);
-        // sprintf(dynName,"%s",name);
-        // routeName = dynName;
-        routeName = nullptr;
-    }
-    int Route::getRouteId() {
-        return routeId;
-    }
-    // bool Route::setRouteName(char* name) {
-    void Route::setRouteName(char* name) {
-        // if (routeName != nullptr) {
-        //     free(routeName);
-        //     routeName=nullptr; 
-        // }
-        // char *dynName;
-        // dynName = (char *) malloc(strlen(name)+1);
-        // // strcpy(dynName, name);
-        // sprintf(dynName,"%s",name);
-
-        // routeName = dynName;
-        // return true;
-        routeName = name;
-    }
-    char* Route::getRouteName() {
-        return routeName;
-    }
-    bool Route::setRouteType(RouteType type) {
-        routeType = type;
-        return true;
-    }
-    RouteType Route::getRouteType() {
-        return (RouteType)routeType;
-    }
-    void Route::setHasReceivedDetails() {
-        hasReceivedDetail = true;
-    }
-    bool Route::getHasReceivedDetails() {
-        return hasReceivedDetail;
-    }
-
-// class TurntableIndex
-
-    TurntableIndex::TurntableIndex(int index, char* name, int angle) {
-        turntableIndexIndex = index;
-        turntableIndexAngle = angle;
-
-        char *dynName;
-        dynName = (char *) malloc(strlen(name)+1);
-        // strcpy(dynName, name);
-        sprintf(dynName,"%s",name);
-
-        turntableIndexName = dynName;
-    }
-    char* TurntableIndex::getTurntableIndexName() {
-        return turntableIndexName;
-    }
-    int TurntableIndex::getTurntableIndexId() {
-        return turntableIndexId;
-    }
-    int TurntableIndex::getTurntableIndexIndex() {
-        return turntableIndexIndex;
-    }
-
-    int TurntableIndex::getTurntableIndexAngle() {
-        return turntableIndexAngle;
-    }
-
-// class Turntable
-
-    Turntable::Turntable(int id, TurntableType type, int position, int indexCount) {
-        turntableId = id;
-        turntableType = type;
-        turntableCurrentPosition = position;
-        turnTableIndexCount = indexCount;
-        turntableName = nullptr;
-        hasReceivedDetail = false;
-        hasReceivedIndexes = false;
-    }
-    int Turntable::getTurntableId() {
-        return turntableId;
-    }
-    void Turntable::setTurntableName(char* name) {
-        turntableName = name;
-    }
-    char* Turntable::getTurntableName() {
-        return turntableName;
-    }
-    bool Turntable::setTurntableType(TurntableType type) {
-        turntableType = type;
-        return true;
-    }
-    TurntableType Turntable::getTurntableType() {
-        return turntableType;
-    }
-    bool Turntable::addTurntableIndex(int index, char* indexName, int indexAngle) {
-        turntableIndexes.add(new TurntableIndex(index, indexName, indexAngle));
-        return true;
-    }
-    bool Turntable::setTurntableCurrentPosition(int index) {
-        if (turntableCurrentPosition != index) {
-            turntableCurrentPosition = index;
-            turntableIsMoving = TurntableMoving;
-            return true;
-        }
-        return false;
-    }    
-    int Turntable::getTurntableCurrentPosition() {
-        return turntableCurrentPosition;
-    }
-    bool Turntable::setTurntableIndexCount(int indexCount) {  // what was listed in the original definition
-        turnTableIndexCount = indexCount;
-        return true;
-    }
-    int Turntable::getTurntableIndexCount() { // what was listed in the original definition
-        return turnTableIndexCount;
-    }
-    int Turntable::getTurntableNumberOfIndexes() { // actual count
-        return turntableIndexes.size();
-    }
-    TurntableIndex* Turntable::getTurntableIndexAt(int positionInLinkedList) {
-        return turntableIndexes.get(positionInLinkedList);
-    }
-    TurntableIndex* Turntable::getTurntableIndex(int indexId) {
-        for (int i=0; i<turntableIndexes.size(); i++) {
-            if (turntableIndexes.get(i)->getTurntableIndexId()==indexId) {
-                return turntableIndexes.get(i);
-            }
-        }
-        return {};
-    }
-    bool Turntable::actionTurntableExternalChange(int index, TurntableState state) {
-        turntableCurrentPosition = index;
-        turntableIsMoving = state;
-        return true;
-    }
-    TurntableState Turntable::getTurntableState() {
-        TurntableState rslt = TurntableStationary;
-        if (turntableIsMoving) {
-            rslt = TurntableMoving;
-        }
-        return rslt;
-    }
-    void Turntable::setHasReceivedDetails() {
-        hasReceivedDetail = true;
-    }
-    bool Turntable::getHasReceivedDetails() {
-        return hasReceivedDetail;
-    }
-
-    void Turntable::setHasReceivedIndexes() {
-        hasReceivedIndexes = true;
-    }
-
-    bool Turntable::getHasReceivedIndexes() {
-        return hasReceivedIndexes;
-    }
